Electron build issues 1. Moved spectron test to a different directory 2. Added a npm command to run spectron test separately
