+++ tests/spectron/jest_spectron.json
@@ -1,0 +1,4 @@
+{
+  "testMatch": ["**/*.spectron.js"],
+  "verbose": true
+}
